BDD tests refactored
