+++ features/steps/commonSteps.py
@@ -1,0 +1,11 @@
+from behave import *
+
+from chronos.models import *
+
+@given('soy empleado')
+def step_impl(context):
+    pass
+
+@given("I am an employee")
+def step_imp(context):
+    pass
